+++ src/app/feature/home/home.module.ts
@@ -5,6 +5,8 @@ import { HomeRoutingModule } from './home-routing.module';
 import { HomeComponent } from './home.component';
 import { FavoriteComponent } from './components/favorite/favorite.component';
 import { CardComponent } from './components/card/card.component';
+import { SharedModule } from '@shared/shared.module';
+import { ModalComponent } from '@shared/components/modal/modal.component';
 
 
 @NgModule({
@@ -15,7 +17,11 @@ import { CardComponent } from './components/card/card.component';
   ],
   imports: [
     CommonModule,
-    HomeRoutingModule
+    HomeRoutingModule,
+    SharedModule
+  ],
+  providers: [
+    ModalComponent
   ]
 })
 export class HomeModule { }