+++ src/app/feature/home/shared/services/character/character.service.ts
@@ -16,4 +16,9 @@ export class CharacterService {
   get(): Observable<Http> {
     return this.http.doGet(`${environment.api}v1/public/characters`);
   }
+
+  getOne(id: number): Observable<Http> {
+    return this.http.doGet(`${environment.api}v1/public/characters/${id}`);
+  }
+
 }