+++ CHANGELOG.md
@@ -14,7 +14,11 @@ straightforward as possible.
 - Added the structure of project
 - Added the module home, this contain the cars of characters and the favorites comics
 - Added the functionality of add random favorites
-- Added the component modal
+- Added the modal component
+- Added the functionality of modal component
+- Added the implementation of modal component in the home module
+  - Added the implementation in card.component
+  - Added the implementation in favorite.component
 - Added the http.service for control the options in the petitions http
 - Added the implementation of http.service in the each one of the services